Move window width detection to client-only code
